read previous session feedback rating back for dialog fix bug w/ authcontroller where it deleted the user node

+++ Droidcon-Boston/app/src/main/java/com/mentalmachines/droidcon_boston/views/rating/RatingRepo.kt
@@ -24,7 +24,7 @@ class RatingRepo(
                     var comments: String = ""
                     data.children.forEach {
                         when (it.key) {
-                            "rating" -> rating = it.value as Int
+                            "rating" -> rating = (it.value as Long).toInt()
                             "feedback" -> comments = it.value as String
                         }
                     }
